Safari-611.1.10/LayoutTests/fast/events/onchange-range-slider-expected.txt

This test verifies that updating the slider for an input element with type=range fires a change event.


PASS Change event fired.
PASS slider.value is "0"
